+++ app/src/main/java/com/github/gotify/messages/provider/MessageImageCombiner.java
@@ -10,7 +10,7 @@ import java.util.concurrent.ConcurrentHashMap;
 public class MessageImageCombiner {
 
     List<MessageWithImage> combine(List<Message> messages, List<Application> applications) {
-        Map<Integer, String> appIdToImage = appIdToImage(applications);
+        Map<Long, String> appIdToImage = appIdToImage(applications);
 
         List<MessageWithImage> result = new ArrayList<>();
 
@@ -26,8 +26,8 @@ public class MessageImageCombiner {
         return result;
     }
 
-    public static Map<Integer, String> appIdToImage(List<Application> applications) {
-        Map<Integer, String> map = new ConcurrentHashMap<>();
+    public static Map<Long, String> appIdToImage(List<Application> applications) {
+        Map<Long, String> map = new ConcurrentHashMap<>();
         for (Application app : applications) {
             map.put(app.getId(), app.getImage());
         }
